+++ authlib/oauth2/rfc7662/__init__.py
@@ -12,5 +12,6 @@
 # flake8: noqa
 
 from .introspection import IntrospectionEndpoint
+from .models import IntrospectionToken
 
-__all__ = ['IntrospectionEndpoint']
+__all__ = ['IntrospectionEndpoint', 'IntrospectionToken']